Cleaned more the code with clang-tidy

+++ jniLibs/Components/src/main/cpp/Components/Samplers/HaltonSeq.cpp
@@ -6,7 +6,7 @@
 
 using Components::HaltonSeq;
 
-HaltonSeq::HaltonSeq(const unsigned long long int domainSize, const unsigned int samples) noexcept :
+HaltonSeq::HaltonSeq(const uint64_t domainSize, const unsigned int samples) noexcept :
         Sampler(domainSize, samples) {
 }
 
@@ -17,7 +17,7 @@ HaltonSeq::HaltonSeq(const unsigned int width, const unsigned int height,
 }
 
 float HaltonSeq::getSample(const unsigned int sample) noexcept {
-    const unsigned long long int current(this->sample_.fetch_add(1ull, std::memory_order_acq_rel));
+    const uint64_t current(this->sample_.fetch_add(1ull, std::memory_order_acq_rel));
     if (current >= (this->domainSize_ * (sample + 1u))) {
         this->sample_.fetch_sub(1ull, std::memory_order_acq_rel);
         return 1.0f;
@@ -26,13 +26,13 @@ float HaltonSeq::getSample(const unsigned int sample) noexcept {
 }
 
 //https://en.wikipedia.org/wiki/Halton_sequence
-float HaltonSeq::haltonSequence(unsigned long long int index, const unsigned int base) noexcept {
+float HaltonSeq::haltonSequence(uint64_t index, const unsigned int base) noexcept {
     float f(1.0f);
     float result(0.0f);
     while (index > 0ull) {
         f /= base;
         result += f * (index % base);
-        index = static_cast<unsigned long long int> (std::floor(index / base));
+        index = static_cast<uint64_t> (std::floor(index / base));
     }
     return result;
 }
