+++ jniLibs/Components/src/main/cpp/Components/Samplers/Random.cpp
@@ -6,11 +6,11 @@
 
 using Components::Random;
 
-Random::Random(const unsigned long long int domainSize, const unsigned int samples) noexcept :
+Random::Random(const uint64_t domainSize, const unsigned int samples) noexcept :
         Sampler(domainSize, samples) {
 }
 
-float Random::getSample(const unsigned int) noexcept {
+float Random::getSample(const unsigned int/*sample*/) noexcept {
     thread_local static std::uniform_real_distribution<float> uniform_dist(0.0f, 1.0f);
     thread_local static std::mt19937 gen(std::random_device{}());
 	//thread_local static std::mt19937 gen(0);
